# project_outline_docs/initial docs/labor-app-openapi-3.0.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Labor Coordination App API",
    "version": "1.0.0",
    "description": "Complete API specification for AV Labor Coordinator event scheduling and invoicing system",
    "contact": {
      "name": "Development Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://api.labor-app.com/api",
      "description": "Production server (future)"
    }
  ],
  "paths": {
    "/events": {
      "get": {
        "operationId": "listEvents",
        "summary": "List all events",
        "tags": ["Events"],
        "responses": {
          "200": {
            "description": "Array of all events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createEvent",
        "summary": "Create a new event",
        "tags": ["Events"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "getEvent",
        "summary": "Get specific event",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "put": {
        "operationId": "updateEvent",
        "summary": "Update event",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteEvent",
        "summary": "Delete event (cascades to requirements and assignments)",
        "tags": ["Events"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/technicians": {
      "get": {
        "operationId": "listTechnicians",
        "summary": "List all technicians",
        "tags": ["Technicians"],
        "responses": {
          "200": {
            "description": "Array of all technicians",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Technician"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTechnician",
        "summary": "Create a new technician",
        "tags": ["Technicians"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicianInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Technician created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Technician"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/technicians/{id}": {
      "get": {
        "operationId": "getTechnician",
        "summary": "Get specific technician",
        "tags": ["Technicians"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Technician found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Technician"
                }
              }
            }
          },
          "404": {
            "description": "Technician not found"
          }
        }
      },
      "put": {
        "operationId": "updateTechnician",
        "summary": "Update technician",
        "tags": ["Technicians"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicianUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Technician updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Technician"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTechnician",
        "summary": "Delete technician",
        "tags": ["Technicians"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Technician deleted"
          }
        }
      }
    },
    "/events/{eventid}/requirements": {
      "get": {
        "operationId": "listRequirements",
        "summary": "List requirements for specific event",
        "tags": ["Requirements"],
        "parameters": [
          {
            "name": "eventid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of requirements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Requirement"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/requirements": {
      "post": {
        "operationId": "createRequirement",
        "summary": "Create a new requirement",
        "tags": ["Requirements"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Requirement created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requirement"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/requirements/{id}": {
      "patch": {
        "operationId": "updateRequirement",
        "summary": "Update requirement",
        "tags": ["Requirements"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Requirement updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requirement"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteRequirement",
        "summary": "Delete requirement (cascades to assignments)",
        "tags": ["Requirements"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requirement deleted"
          }
        }
      }
    },
    "/assignments": {
      "get": {
        "operationId": "listAssignments",
        "summary": "List all assignments",
        "tags": ["Assignments"],
        "responses": {
          "200": {
            "description": "Array of all assignments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Assignment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAssignment",
        "summary": "Create a new assignment (CRITICAL - requires all 25 fields)",
        "tags": ["Assignments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assignment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input - must include all 25 fields"
          },
          "500": {
            "description": "Missing required fields"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "clientname": {
            "type": "string"
          },
          "clientcontact": {
            "type": "string"
          },
          "clientphone": {
            "type": "string"
          },
          "clientemail": {
            "type": "string"
          },
          "clientaddress": {
            "type": "string"
          },
          "ponumber": {
            "type": "string"
          },
          "startdate": {
            "type": "string",
            "format": "date"
          },
          "enddate": {
            "type": "string",
            "format": "date"
          },
          "totaltechpayout": {
            "type": "number",
            "format": "double"
          },
          "totallaborcost": {
            "type": "number",
            "format": "double"
          },
          "totalcustomerbilling": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["id", "name"]
      },
...
    }
  }
}